refactor(material-ui) : merge material-ui and material-ui-dashboard
