+++ build.sbt
@@ -178,29 +178,13 @@ lazy val `material-ui` =
       webpackDevServerPort := 8008,
       stFlavour := Flavour.Slinky,
       stExperimentalEnableImplicitOps := true,
-      stIgnore ++= List("@material-ui/icons"),
-      Compile / npmDependencies ++= Seq(
-        "@material-ui/core" -> "3.9.3", // note: version 4 is not supported yet
-        "@material-ui/styles" -> "3.0.0-alpha.10" // note: version 4 is not supported yet
-      )
-    )
-
-lazy val `material-ui-dashboard` =
-  project
-    .enablePlugins(ScalablyTypedConverterPlugin)
-    .configure(baseSettings, browserProject, reactNpmDeps, bundlerSettings)
-    .settings(
-      useYarn := true,
-      webpackDevServerPort := 8008,
-      stFlavour := Flavour.Slinky,
-      stExperimentalEnableImplicitOps := true,
       Compile / npmDependencies ++= Seq(
         "@material-ui/core" -> "3.9.3", // note: version 4 is not supported yet
         "@material-ui/styles" -> "3.0.0-alpha.10", // note: version 4 is not supported yet
-        "@material-ui/icons"   -> "3.0.2",
-        "recharts"           -> "1.8.5",
+        "@material-ui/icons" -> "3.0.2",
+        "recharts" -> "1.8.5",
         "@types/recharts" -> "1.8.10",
-        "@types/classnames"   -> "2.2.10",
+        "@types/classnames" -> "2.2.10",
       )
     )
