--- material-ui-dashboard/src/main/js/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Material-ui demo</title>
-</head>
-<body>
-<div id="container"></div>
-
-<script type="text/javascript" src="material-ui-dashboard-fastopt-bundle.js"></script>
-</body>
-</html>
